svn merge -r16727:16742 https://eris.llnl.gov/svn/slurm/branches/slurm-1.3

+++ trunk/src/plugins/proctrack/rms/proctrack_rms.c
@@ -94,7 +94,7 @@ extern int slurm_container_create (slurmd_job_t *job)
 	/*
 	 * Return a handle to an existing prgid or create a new one
 	 */
-	if (rms_getprgid (job->jmgr_pid, &prgid) < 0) {
+	if (rms_getprgid ((int) job->jmgr_pid, &prgid) < 0) {
 		int fd = _prg_destructor_fork();
 		/* Use slurmd job-step manager's pid as a unique identifier */
 		prgid = job->jmgr_pid;
